refactor: extract OpportunityWindow from AnalyticsState
